src/unit.py
```python
class Unit:
    def __init__(self, unit_type, health, damage, speed):
        self.unit_type = unit_type
        self.health = health
        self.damage = damage
        self.speed = speed

    def move(self, lane):
        # Logic for unit movement along the lane
        pass

    def attack(self, target):
        # Logic for unit attacking a target
        pass

    def take_damage(self, damage):
        # Logic for unit taking damage
        pass

class UnitWave:
    def __init__(self, units, spawn_interval):
        self.units = units
        self.spawn_interval = spawn_interval

    def spawn_units(self):
        # Logic for spawning units at intervals
        pass

class AugmentedUnit(Unit):
    def __init__(self, unit_type, health, damage, speed, augment):
        super().__init__(unit_type, health, damage, speed)
        self.augment = augment

    def apply_augment(self):
        # Logic for applying augment to unit
        pass
```