tests/unit_tests.py
```python
import unittest
from src import game, map, base, tower, unit, resource, champion, fog_of_war, augment_system

class TestGameMechanics(unittest.TestCase):

    def setUp(self):
        self.game_state = game.GameState()
        self.player_resources = resource.Resources()
        self.augments = augment_system.Augments()

    def test_map_layout(self):
        map_layout = map.Map()
        self.assertIsNotNone(map_layout)
        self.assertEqual(len(map_layout.lanes), 3)

    def test_base_building(self):
        player_base = base.Base()
        self.assertIsNotNone(player_base)
        self.assertEqual(player_base.health, 100)

    def test_tower_placement(self):
        tower_placement = tower.Tower()
        self.assertIsNotNone(tower_placement)
        self.assertEqual(tower_placement.damage, 10)

    def test_unit_waves(self):
        unit_wave = unit.Unit()
        self.assertIsNotNone(unit_wave)
        self.assertEqual(unit_wave.health, 5)

    def test_resource_management(self):
        resource_management = resource.Resources()
        self.assertIsNotNone(resource_management)
        self.assertEqual(resource_management.gold, 100)

    def test_champion_control(self):
        champion_control = champion.Champion()
        self.assertIsNotNone(champion_control)
        self.assertEqual(champion_control.health, 50)

    def test_fog_of_war(self):
        fog_of_war_system = fog_of_war.FogOfWar()
        self.assertIsNotNone(fog_of_war_system)
        self.assertEqual(fog_of_war_system.visibility, False)

    def test_augment_system(self):
        augment_system_test = augment_system.Augments()
        self.assertIsNotNone(augment_system_test)
        self.assertEqual(augment_system_test.power, 0)

if __name__ == '__main__':
    unittest.main()
```